Updating readme.md file
Note that on line 11, the test results are being shown after running the live server for test.html file and not index.html file. 
Please check if I am right or not.
Also, if I am wrong then please specify what we have to do after running the index.html as there are three links of which none are properly functional.
Thank you!